Add "My Profile" Page Make integer parsing in ConfigService more robust Add validation to config pages Implement Cron expression validator

+++ src/main/kotlin/de/grimsi/gameyfin/config/ConfigService.kt
@@ -141,7 +141,7 @@ class ConfigService(
         return when (configProperty.type) {
             String::class -> value as T
             Boolean::class -> value.toBoolean() as T
-            Int::class -> value.toInt() as T
+            Int::class -> value.toFloat().toInt() as T
             Float::class -> value.toFloat() as T
             else -> {
                 throw RuntimeException("Unknown config type ${configProperty.type}: '$value' for key ${configProperty.key}")
